requirements.txt
mysql-connector-python==8.2.0
PyYAML==6.0.1
python-dateutil==2.8.2
rapidfuzz==3.6.1
openai>=1.40.0,<2.0.0
parsera==0.1.7
requests==2.31.0
playwright==1.57.0
beautifulsoup4==4.12.2